+++ lib/src/grains/dynamic/mod.rs
@@ -19,13 +19,13 @@ pub struct Granulator {
   next_grain: usize,
   grains: Vec<Grain>,
 
-  samplerate: f32,
+  samplerate: u32,
   sr_recip: f32,
 }
 
 
 impl Granulator {
-  pub fn new(envelope: Vec<f32>, samplerate: f32, num_grains: usize, buf_size: usize) -> Self {
+  pub fn new(envelope: Vec<f32>, samplerate: u32, num_grains: usize, buf_size: usize) -> Self {
     // Buffer to hold recorded audio
     let buffer = vec![0.0; buf_size];
 
@@ -48,7 +48,7 @@ impl Granulator {
       recording: false,
       next_grain: 0,
       samplerate,
-      sr_recip: 1.0 / samplerate,
+      sr_recip: 1.0 / samplerate as f32,
     }
   }
 
@@ -124,9 +124,9 @@ impl GrainTrait for Granulator {
     self.envelope = shape;
   }
 
-  fn set_samplerate(&mut self, samplerate: f32) {
+  fn set_samplerate(&mut self, samplerate: u32) {
     self.samplerate = samplerate;
-    self.sr_recip = 1.0 / samplerate;
+    self.sr_recip = 1.0 / samplerate as f32;
   }
